Add update command
