+++ cmd/update/runner.go
@@ -1,4 +1,4 @@
-package client
+package update
 
 import (
 	"context"
@@ -33,5 +33,10 @@ func (r *runner) Run(cmd *cobra.Command, args []string) error {
 }
 
 func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) error {
+	err := cmd.Help()
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
 	return nil
 }
